pyproject.toml
[project]
name = "subdog"
version = "0.1.0"
description = "Async subdomain resolver with concurrency + rate limiting"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "dnspython>=2.8.0",
    "tqdm>=4.67.3",
    "trio>=0.33.0",
]

[project.scripts]
subdog = "main:main"

[tool.ty.environment]
python = ".venv"
